Continue moving datamanagement to plugin. - Move FiberData from medVtkInria to medVtkFibersData plugin.
