+++ src/plugins/legacy/medVtkFibersData/CMakeLists.txt
@@ -47,6 +47,7 @@ list_source_files(${TARGET_NAME}
   . #current dir
   datas
   interactors
+  manager
   readers
   writers
   )
