+++ src/plugins/legacy/medVtkFibersData/vtkFiberDataSet.h
@@ -14,7 +14,7 @@
 #pragma once
 
 #include <vtkMultiBlockDataSet.h>
-#include <medVtkInriaExport.h>
+#include <medVtkFibersDataPluginExport.h>
 #include <map>
 
 class vtkPolyData;
@@ -24,7 +24,7 @@ class vtkPolyData;
    as vtkPolyData as well.
  */
 
-class MEDVTKINRIA_EXPORT vtkFiberDataSet : public vtkMultiBlockDataSet
+class MEDVTKFIBERSDATAPLUGIN_EXPORT vtkFiberDataSet : public vtkMultiBlockDataSet
 {
  public:
   static vtkFiberDataSet *New();
